pyproject.toml
[project]
name = "books-ml-ordered"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "pydantic-settings>=2.9.1",
    "typer>=0.16.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]

[tool.pytest.ini_options]
pythonpath = [
  "src",
  "tests"
]
markers = [
  "acceptance: acceptance tests only on local"
]
